news_collector/requirements.txt
beautifulsoup4>=4.12.0
feedparser>=6.0.8
requests>=2.28.0
tqdm>=4.64.0
warcio>=1.7.0
newspaper3k>=0.2.8
lxml>=4.9.0
matplotlib>=3.7.0
pandas>=2.0.0
pyarrow>=14.0.0
bs4>=0.0.1 